+++ blocklets/ai-studio/src/app.tsx
@@ -1,5 +1,6 @@
 import { LocaleProvider } from '@arcblock/ux/lib/Locale/context';
 import { ToastProvider } from '@arcblock/ux/lib/Toast';
+import withTracker from '@arcblock/ux/lib/withTracker';
 import { SubscribeButton } from '@blocklet/ai-kit/components';
 import { Dashboard } from '@blocklet/studio-ui';
 import Footer from '@blocklet/ui-react/lib/Footer';
@@ -7,6 +8,7 @@ import { Box, CssBaseline, GlobalStyles, ThemeProvider } from '@mui/material';
 import { ReactNode, Suspense, lazy } from 'react';
 import {
   Navigate,
+  Outlet,
   Route,
   RouterProvider,
   createBrowserRouter,
@@ -18,6 +20,7 @@ import { RecoilRoot } from 'recoil';
 import AigneLogo from './components/aigne-logo';
 import ErrorBoundary from './components/error/error-boundary';
 import Loading from './components/loading';
+import { PlanUpgrade } from './components/multi-tenant-restriction';
 import { SessionProvider, useInitialized, useIsPromptEditor, useIsRole } from './contexts/session';
 import { translations } from './locales';
 import { theme } from './theme/theme';
@@ -92,6 +95,7 @@ export default function App() {
               <SessionProvider serviceHost={basename} protectedRoutes={[`${basename}/*`]}>
                 <Suspense fallback={<Loading fixed />}>
                   <ErrorBoundary>
+                    <PlanUpgrade />
                     <AppRoutes basename={basename} />
                   </ErrorBoundary>
                 </Suspense>
@@ -104,6 +108,12 @@ export default function App() {
   );
 }
 
+function Root() {
+  return <Outlet />;
+}
+
+const TrackedRoot = withTracker(Root);
+
 function AppRoutes({ basename }: { basename: string }) {
   const initialized = useInitialized();
   const isPromptEditor = useIsPromptEditor();
@@ -113,7 +123,7 @@ function AppRoutes({ basename }: { basename: string }) {
 
   const router = createBrowserRouter(
     createRoutesFromElements(
-      <Route path="/" ErrorBoundary={RouterErrorBoundary}>
+      <Route path="/" element={<TrackedRoot />} ErrorBoundary={RouterErrorBoundary}>
         <Route index element={isPromptEditor ? <Navigate to="projects" replace /> : <Home />} />
         <Route path="explore/*" element={<ExploreCategory />} />
         <Route path="apps/:appId" element={<AppPage />} />