+++ blocklets/ai-studio/src/locales/en.ts
@@ -857,4 +857,18 @@ export default flatten({
   selectEntryAgentTitle: 'Select Entry Agent',
   selectEntryAgentDescription: 'Before deployment, you need to set up the entry point agent.',
   activeWhen: 'Active when',
+  multiTenantRestriction: {
+    requestLimit: {
+      desc: 'You have reached the maximum request limit in this project. You may increase the limit by upgrading your plan to Premium.',
+    },
+    projectLimit: {
+      desc: 'You have reached the maximum project limit. Upgrade your plan to Premium to manage more projects.',
+    },
+    customBrand: {
+      desc: 'Branding customization is available only in the Premium plan or Serverless AIGNE. Please upgrade to the Premium plan to access this feature.',
+    },
+    cronJobs: {
+      desc: 'The Cron Jobs feature is available only in Serverless AIGNE. Please run a Serverless AIGNE to access this feature.',
+    },
+  },
 });