+++ packages/ai-runtime/src/front/runtime-components/SimplePage/index.tsx
@@ -6,57 +6,55 @@ import { AgentErrorView } from '../../components/AgentErrorBoundary';
 import CustomComponentRenderer from '../../components/CustomComponentRenderer/CustomComponentRenderer';
 import SimpleLayout from '../../components/Layout/SimpleLayout';
 import { DEFAULT_HEADER_COMPONENT_ID } from '../../constants';
+import { useActiveAgent } from '../../contexts/ActiveAgent';
 import { ComponentPreferencesBase, ComponentPreferencesProvider } from '../../contexts/ComponentPreferences';
 import { CurrentAgentProvider, useCurrentAgent } from '../../contexts/CurrentAgent';
 import { CurrentMessageProvider } from '../../contexts/CurrentMessage';
 import { MessageItem, useSession } from '../../contexts/Session';
 import { useAppearances, useProfile } from '../../hooks/use-appearances';
+import InputsView from '../SimpleChat/InputsView';
+import OpeningMessageView from './OpeningMessageView';
 
-export interface SimplePagePreferences extends ComponentPreferencesBase {}
+export interface SimplePagePreferences extends ComponentPreferencesBase {
+  resultTitle?: string;
+}
 
-export default function SimplePage({ resultTitle, ...preferences }: { resultTitle?: string } & SimplePagePreferences) {
+export default function SimplePage({ ...preferences }: SimplePagePreferences) {
+  const { aid: activeAid } = useActiveAgent();
+
   return (
     <ComponentPreferencesProvider {...preferences}>
       <SimpleLayout pb={4}>
-        <HeaderView />
+        <CurrentAgentProvider aid={activeAid}>
+          <HeaderView px={{ xs: 2, sm: 3 }} />
 
-        <InputView className="aigne-inputs aigne-simple-page-inputs" />
+          <OpeningMessageView my={4} px={{ xs: 2, sm: 3 }} />
 
-        <OutputView className="aigne-outputs aigne-simple-page-outputs" resultTitle={resultTitle} />
+          <InputsView className="aigne-inputs aigne-simple-page-inputs" />
+
+          <OutputView
+            className="aigne-outputs aigne-simple-page-outputs"
+            resultTitle={preferences.resultTitle}
+            px={{ xs: 2, sm: 3 }}
+          />
+        </CurrentAgentProvider>
       </SimpleLayout>
     </ComponentPreferencesProvider>
   );
 }
 
-function HeaderView() {
+function HeaderView(props: StackProps) {
   const { aid } = useCurrentAgent();
   const profile = useProfile({ aid });
 
   return (
-    <Stack>
+    <Stack {...props}>
       <CustomComponentRenderer
         aid={aid}
         output={profile.outputSettings}
         componentId={profile.appearance.componentId || DEFAULT_HEADER_COMPONENT_ID}
         properties={profile.appearance.componentProperties}
         props={profile.appearance.componentProps}
-      />
-    </Stack>
-  );
-}
-
-function InputView({ ...props }: StackProps) {
-  const { aid } = useCurrentAgent();
-  const { appearanceInput } = useAppearances();
-
-  return (
-    <Stack {...props}>
-      <CustomComponentRenderer
-        aid={aid}
-        output={appearanceInput.outputSettings}
-        componentId={appearanceInput.componentId}
-        properties={appearanceInput.componentProperties}
-        props={appearanceInput.componentProps}
       />
     </Stack>
   );
@@ -94,7 +92,6 @@ const OutputItemView = memo(({ message }: { message: MessageItem }) => {
           <CustomComponentRenderer
             aid={message.aid}
             output={appearanceOutput.outputSettings}
-            key={message.id}
             componentId={appearanceOutput.componentId}
             properties={appearanceOutput.componentProperties}
             props={appearanceOutput.componentProps}