+++ packages/aigne-sdk/src/components/AgentView/AgentView.tsx
@@ -1,24 +1,6 @@
-import { CustomComponentRenderer } from '@blocklet/pages-kit/components';
+import { Runtime } from '@blocklet/ai-runtime/front';
+import { ComponentProps } from 'react';
 
-import { AIGNE_RUNTIME_CUSTOM_COMPONENT_ID } from '../../constants';
-import { ComponentPreferencesBase, ComponentPreferencesProvider } from '../ai-runtime';
-
-export default function AgentView({
-  blockletDid,
-  aid,
-  working,
-  ...props
-}: {
-  blockletDid?: string;
-  aid: string;
-  working?: boolean;
-} & ComponentPreferencesBase & { [key: string]: any }) {
-  return (
-    <ComponentPreferencesProvider {...props}>
-      <CustomComponentRenderer
-        componentId={AIGNE_RUNTIME_CUSTOM_COMPONENT_ID}
-        props={{ blockletDid, aid, working, props }}
-      />
-    </ComponentPreferencesProvider>
-  );
+export default function AgentView(props: ComponentProps<typeof Runtime>) {
+  return <Runtime {...props} />;
 }